+++ aten/src/ATen/core/Macros.h
@@ -45,6 +45,21 @@
   classname(const classname&) = delete;       \
   classname& operator=(const classname&) = delete
 
+
+#if defined(__ANDROID__)
+#define AT_ANDROID 1
+#define AT_MOBILE 1
+#elif (defined(__APPLE__) &&                                            \
+       (TARGET_IPHONE_SIMULATOR || TARGET_OS_SIMULATOR || TARGET_OS_IPHONE))
+#define AT_IOS 1
+#define AT_MOBILE 1
+#elif (defined(__APPLE__) && TARGET_OS_MAC)
+#define AT_IOS 1
+#define AT_MOBILE 0
+#else
+#define AT_MOBILE 0
+#endif // ANDROID / IOS / MACOS
+
 namespace at {
 inline int stoi(const std::string& str) {
 #if defined(__ANDROID__)
