+++ aten/src/ATen/core/OptionsGuard.cpp
@@ -1,7 +1,9 @@
-#include <ATen/OptionsGuard.h>
+#include <ATen/core/OptionsGuard.h>
 #include <ATen/core/optional.h>
 
 namespace at {
+
+#if !AT_MOBILE && !defined(CAFFE2_FB_LIMITED_MOBILE_CAPABILITY)
 
 thread_local at::optional<TensorOptions> DefaultTensorOptions::options_;
 
@@ -12,5 +14,15 @@ TensorOptions& DefaultTensorOptions::get() {
   }
   return *options_;
 }
+
+#else
+
+TensorOptions DefaultTensorOptions::options_(/*use_thread_local_default_options=*/false);
+
+const TensorOptions& DefaultTensorOptions::get() {
+  return options_;
+}
+
+#endif
 
 } // namespace at
