+++ aten/src/ATen/core/TensorOptions.cpp
@@ -1,9 +1,9 @@
 #include <ATen/core/TensorOptions.h>
 
-#include <ATen/Device.h>
+#include <ATen/core/Device.h>
 #include <ATen/core/Layout.h>
-#include <ATen/OptionsGuard.h>
-#include <ATen/ScalarType.h>
+#include <ATen/core/OptionsGuard.h>
+#include <ATen/core/ScalarType.h>
 #include <ATen/core/optional.h>
 #include <ATen/core/ScalarType.h>
